goldenkey/requirements.txt
APScheduler==3.8.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.10.0
django-polymorphic==3.0.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
filelock==3.4.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
oauthlib==3.1.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
virtualenv==20.10.0
zipp==3.6.0
